[FEATURE] Add cache:flush command
A cache:flush CLI command is added that provides
lowlevel cache flush functionality as in installtool.

It is made extra sure that the command will be executable,
even if all caches are stale or even broken. That means this
command is able to repair broken di, localconf and tca caches.
Technically this became possible due to preparations for
lowlevel cli commands in #86248.

Releases: master
Resolves: #90197
Change-Id: I920867fa8a1ae908acf9788f5de2c6c79eb452d5