[FEATURE] Support PHP 8 style Channel attribute for logger injection
Services are now able to control the component name that an
injected logger is created with.
This allows to group logs of related classes and is basically
a channel system as often used in monolog.

Therefore and as prepration for future monolog logging,
the attribute is named Channel.

The Channel attribute is supported for constructor argument
injection as a parameter attribute and for LoggerAwareInterface
DI services as a class attribute.

This feature is only available with PHP 8.
The channel attribute will be gracefully ignored in PHP 7,
and the classic component name will be used instead.

Releases: master
Resolves: #95079
Related: #95044
Change-Id: I4b710677bc20c4200fc14b299865f4c7f95e4d0d