[WIP][BUGFIX] Display root page in pagetree always
This applies to rendering of database mounts (e.g. from groups or
from workspaces). They are now shown as children of the virtual
root page (uid=0).

This is to allow workspace users to get an overview of all
changes of all mounts by navigating to the virtual page uid=0.
pagetree has been refactored to a new controller.

TODO:
 * The git diff is huge and hard to review…
 * Too intrusive for v10?
 * Split/rename functions? (BC?)
 * test test test…
 * The current code reads like it was kind-of desired
   to hide pid=0 for cases where users have db-mounts.

   Maybe discard this change and rather provide
   an "overall" overview via the backend modules itself?

Releases: master, 10.4
Resolves: #95854
Change-Id: I7f6370f327711396193cf56b63f15876350c2559