[TASK] Reenable taboverride in grunt copy action
Instead of disabling the grunt copy,
because of a hardcoded sourcemap reference in the
minified build, we simply use the non
standard build instead and minify the file
ourselves. This results in 25 bytes increased
file size, but that's fine as it allows to auto-update
this library.

While at it, the filename is changed to
reflect the npm module name.

Resolves: #96324
Related: #88999
Related: #96323
Releases: main
Change-Id: I06c359b3600a16732568877831d1b719f21ffa9a