[TASK] Use vanilla script-loading for ckeditor
CKEditor v4 is not available as ES6 module and
can not be easily transformed into a strict
module compatible module (as implied by
<script type="module">): CKEditor requires
global `this` value (this = window)
and thus doesn't work in strict mode.

Therefore CKEditor v4 is now loaded
via a plain <script async> tag in
preparation for requirejs to ES6 module
transition. (It hasn't been an AMD
module either, but was shimed
by requirejs to be loadable as
AMD module).

Change-Id: Ife0b476ba29b85f85dfd654ea096cdf30c47b6ef