[BUGFIX] Fix ordering of db mounts in page tree
With the changes in #95854, db mounts (e.g for editors)
are no longer sorted by db-mount-definition,
but by sorting of the mountpoints.

DB mounts aren't sorted against each other,
that means sorting values are not meaningful, and
results in uncontrolable orderings.

Avoid groupAndSortPages() in that case
and simply preserve the entryPoint input-array
ordering by using array_values. Also revert
groupAndSortPages() to the state prior to #95854,
as the forcePid parameter would now be unused.

Releases: main, 11.5, 10.4
Resolves: #96359
Related: #95854
Change-Id: I64b8d90c68246cc0ac01abcbf3678aa1055189e3