[BUGFIX] Avoid duplicate pages in pagetree when rootPid is mounted
With the changes in #95854, implicit rootPid
dbmounts (e.g. when an admin user as additional
pages mounted as webmount in the user settings),
caused duplicated pages in pagetree, because
pages with pid=0 where fetched multiple times.

Resort to classic entryPoint handling
when the rootPid (0) is included as webmount.

Releases: main, 11.5, 10.4
Resolves: #96367
Related: #95854
Change-Id: I7f421995f2d4aa88bb266fec684565102a4dc080