[TASK] Tag importmaps to include configuration for dynamic imports
Dynamic JavaScript imports require their respective importmap
configurations to be loaded, therefore it is now allowed to tag
import configurations in order to be loaded when a certain
dispatcher is added to the current response.

This patch adds tags for currently used dynamic imports.

Note: In the pageRenderer->loadRequireJsModule we currently
transparently load all configurations in backend context, but
once we switch all these calls to loadJavaScriptModule(),
we need to be very explicit about all the required
importmaps. This patch prepares for that.

Resolves: #96710
Related: #96323
Releases: main
Change-Id: I7b9eb0cb0a26b90ae217435c15c04238ffe4baa2