[WIP][TASK] Tag importmaps to include configuration for dynamic imports
Dynamic JavaScript imports require their respective importmap
configurations to be loaded, therefore it is now allowed to tag
import configurations in order to be loaded when a certain
dispatcher is added to the current response.

Add tags for currently used dynamic imports.

Note: In the pageRenderer->loadRequireJsModule we currently
transparently load all configurations in backend context,
once we switch all these calls to loadJavaScriptModule(),
we need to pre very explicit about all the required
importmaps. This patch prepares for that.

TODO:

 * Do want want tags at all? or drop all this depenency stuff
   in the backend? but then, how to avoid extension disclosure
   in the frontend and login screen?
 * Should tags be used explicitly instead of adding an event
   to ImportMap in order to listen for imported modules?
 * Drop resolution from ResolveJavaScriptImportEvent? (not used in any
   core module, added because it may be useful to others)
 * Split PageRenderer serialization fixes into a separate patch
 * Support proper JavaScriptRenderer serialization via
   updateState()

Resolves: #
Related: #96323
Releases: main
Change-Id: I7b9eb0cb0a26b90ae217435c15c04238ffe4baa2