[POC][!!!][FEATURE] Introduce CSS variables for FSC styling options
This patch introduces css variables for fluid styled
content TypoScript styling options.

The advantage is that themes can make use of those
css variables. Before, there was no other way to
use style settings in own css stylesheets.

The FSC stylesheets are now generated in the main
build process, by extracting them in own scss files.
The resulting minified css files are included by the
AssetCollector in their corresponding template.

The styles still respect config.removeDefaultCss and
are not included if set to "1". Also, if the
optional TypoScript template for CSS is not included
the styles won't be added.

Resolves: #96939
Releases: main
Change-Id: I68c8496b4a3403e20e6dbee791a2fb5398a15a2b