[FEATURE] Refresh the look of pagemodule
The patch simplifies and optimizes the readability of the page module
interface. It is based on the UX Team's concept of improving user
experience for editors:
https://typo3.org/article/structured-content-initiative-feedback-wanted

The underlying CSS was refactored and optimised for future adaptations.
The content element boxes and their header buttons are visually
simplified. The hidden content elements are now differenciated with a
better opacity and a dotted border. The New Content buttons were placed
in the centre in preparation for implementation of later concepts.
A new button for the content element context menu is added in the
content element header right button bar, in place of the delete button.
Based on common UI concepts, the delete button has been moved to
the context menu.

This patch was initiated by the UX Team and implemented by Benjamin
Kott.

Resolves: #97306
Resolves: #95247
Related: #94629
Releases: main
Change-Id: I9b4e21a7fc68ebcb17d18f65bf00096f3a295cfb