[BUGFIX] Fix page reloading on click on preview button
If a preview page is already open, clicking on the "view" button in the
content editor window would not reload the page. That means, that the
preview page would not show changes if the preview page url and the
target url are similar (e.g. only differentiate in hash/query or are
exactly the same).

This patch looks if a window with the given windowName is already open
(looking in `this.windows` did not work well because the existing
windows aren't saved in a global context, yet) and compares the URL of
an open window (if there's one) with the given URL. If the URLs are
similiar (i.e. the origin and the path name are the same) the page will
be reloaded.

Resolves: #93706
Releases: main, 11.5
Change-Id: Ib3d9e49f56e6ef45df1cd7d82fa58dffb74a3750