[!!!][TASK] Hash package version into public resource paths
This applies to installations in TYPO3 composer mode using
typo3/cms-composer-installers v5 where TYPO3 extensions
are stored in vendor/ and their public resources are linked
to public/_assets/.

The symlink name in public/_assets/ is now extended to be not
just a hash of the relative path to the vendor folder, but also
acts as a cacheHash, by including a random cache hash into the
symlink name.

That means an extension update of packages/extensions
(or a `composer dump-autoload`) will trigger a new symlink name.
This has the drawback that names in public/_assets/ are
no longer predictable, but brings the benefit that this will allow
TYPO3 to skip modification-timstamp-lookups in
GeneralUtility::createVersionNumberedFilename() and to generate
performant JavaScript importmaps via trailing slashes[1] syntax.

[1] https://github.com/WICG/import-maps#packages-via-trailing-slashes

Releases: main
Resolves: #98481
Related: #96323
Change-Id: I132ea343c12719bd9e6f24fe82915e2c4508faef