[TASK] Render JavaScript modules as script tag if possible
With the removal of es-module-shims in #101372 it is possible[1]
to add a fastpath for JavaScript modules that do not require
initializer methods to be called externally.

The fastpath is activated when all modules match the
requirements, in order to avoid mixing explicit module script
tags and @typo3/core/java-script-item-handler.js invocations.

This allows the browser to load the set of modules from the list
of <script type="module"> tags while global configuration is written
as plain JSON and the intermediate module java-script-item-handler.js
is skipped.

Note that the litNonce is migrated to use global assignment
infrastructure to be guaranteed to be rendered after the importmap
in order to prevent occasional false-positive errors in Firefox 115
> Uncaught TypeError: The specifier "@typo3/backend/modal.js" was
> a bare specifier, but was not remapped to anything.
(The backend and all actions would have worked, as all scripts are
loaded despite of these errors, but javascript console would be
flooded with these message. This is fixed in Firefox 116 and
was probably triggered because `<script type="importmap">` has not
been placed as `<script>` element on the page.

[1] we do not longer need to care about missing support for
    `import()` expressions in `<script type="module" src="…">`
    tags, which were only supported via a fail-first algorithm
    in es-module-shims, and have therefore not been used

Releases: main
Resolves: #101297
Related: #96323
Change-Id: If8d29741081a52508a8981ea0f5bb098c2fb808f