[BUGFIX] Inject logger into YamlFileLoader
YamlFileLoader needs to be instantiated via `self::new` to
ensure that a logger instance is injected.

The issues was introduced with #103187 where YamlFileLoader was added to
be provided by the EXT:install ServiceProvider. This is actually not
quite right, as a service is always to be provided by the extension that
ships the service, not by the consumer.

Therefore the YamlFileLoader service definition is now moved into
EXT:core.

Resolves: #103444
Related: #103187
Releases: main
Change-Id: Ifa3b690a43239c26903443789f59fcd2768fd9d5