[TASK] Implement TypoScript provider for profiles and sites
TYPO3 sites are enhanced to be able to operate as TypoScript
templates. They act similar to sys_template records with
"clear" and "root" flags set. By design a site always defines
a new scope (root-flag) and does not inherit from parent sites
in the rootline (clear-flag). This is not configurable.
TypoScript code sharing is to be implemented via sharable profiles.

Dependencies can be included via (recursive) profiles (#103437).
This mechanism supersedes the previous static_file_include's
or manual `@import` statements (they are still fine for local includes,
but should be avoided for cross-profile/extensions dependencies),
as profiles are automatically ordered and deduplicated.

Resolves: #103439
Releases: main
Change-Id: I971743fc551e51d945f45335dc6ad76404c6edba