[WIP][FEATURE] TYPO3 CMS Content Blocks
Backend
=======

- Loading of Content Block definitions which reside in extensions
- Parsing and compilation of the YAML files into data objects
- Code generation (mainly TCA, SQL and page TS)
- Caching

Frontend
========

- Generation of TypoScript for frontend rendering definition
- Decoration of Fluid data variable (ContentBlockData)
- Specialised ViewHelper for resources inside ContentBlock folder

BE and FE
=========

- Relation Resolving based on RelationHandler (Also for BE previews)
- Publishing of Assets folder

@todo Right now work is being done in the frontend part, which might
replace the Content Block implementations. The focus should be to
extract as much as possible to standalone patches. E.g. the Relation
resolving as a general purpose tool.

@todo Better solution for asset publishing.

N

This patch introduces the concept of a "Content Block" as a first class
citizen into the TYPO3 core.

A Content Block defines exactly one "Content Type". A YAML file serves
as a basis for this definition. The main concept is to encapsulate all
assets which are needed to define such a type. These assets are TCA,
SQL, XLF and icons for previews. Having this component-based approach
it is now easy to share and reuse single blocks in other projects.

Content Blocks itself is more or less a loader/compiler/generator type
of extension. It is based on already existing functionality and just
adds another abstraction layer.

The structure of a Content Block consists of the EditorInterface.yaml
file, the "Assets" folder with icons and the "Source" folder, which may
contain xlf translations. Only the YAML file is required.

   ├── Assets
   │   └── Icon.svg
   ├── Source
   │   └── Language
   │       └── Labels.xlf
   └── EditorInterface.yaml

Single Content Blocks need a host extension to live inside. Depending
on the folder they are placed in, they are interpreted as Content
Elements, Page Types or generic Record Types.

   ├── Classes
   ├── Configuration
   ├── ContentBlocks
   │   ├── ContentElements
   │   │   ├── block1
   │   │   └── block2
   │   ├── PageTypes
   │   │   ├── block3
   │   │   └── block4
   │   └── RecordTypes
   │       ├── block5
   │       └── block6
   ├── Resources
   └── composer.json

The YAML definition defines a project-wide, unique identifier. It also
defines all fields and their position, in which they should appear in
the editing interface (FormEngine):

    name: vendor/content-block-name
    fields:
      - identifier: my_text_field
        type: Text

Commands executed:
  composer require symfony/var-exporter:^7.1

Resolves: #98135

Change-Id: I2c20f1a4f4ca4e9ff6bce8b2c4741e1c577dc9ca