[BUGFIX] Fix detection of local settings in settings overview
Site local settings (defined in the sites settings.yaml file)
need to be taken into account in the site overview, only
if they are based on definitions. Therefore the setting
map (which is based on definitions) has to be evaluated
instead of the flat settings map (which contains all
settings, defined and legacy anonymous settings).

Releases: main
Resolves: #104886
Related: #104794
Change-Id: I49f8d01c3947302ff75a6aac0e3e64cdeca819c8