[WIP][!!!][FEATURE] Support unqualified ISO8601 in DataHandler instead of fake UTC0
Align DataHandler datetime handling to accept unqualified ISO8601
(also called LOCALTIME [1][2][3]) and interpret those strings in
server localtime and adapt the TYPO3 backend to use ISO8601
LOCALTIME as interchange format between FormEngine PHP and
HTML/JS code.

Why localtime without offset information (=unqualified)?
--------------------------------------------------------

First and foremost: Conceptionally that's not a change, rather
a technical cleanup of the current approach as we have been effectively
using ISO8601 localtime as interchange format between FormEngine and
DataHandler:

Currently a faked ISO8601 UTC0 string is used that is interpreted as
localtime. For example 1999-11-11T11:11:11Z on a server timezone
Europe/Berlin (offset +01:00 in CET) is interpreted as:
1999-11-11T11:11:11+01:00
           ^^

Formally seen that's wrong as 1999-11-11T11:11:11Z actually refers to
the 12th hour in germany:
1999-11-11T12:11:11+01:00
           ^^

Why not switch to real UTC + offsets directly?
----------------------------------------------

Editors (currently) do only enter dates, no timezone information.
They are supposed to enter dates in the servers timezone (which is
currently not introspectable from the backend UI, but that's another
topic and an UX issue).

To transform the raw input into the server timezone it's important
to know what the *intended* value was, that means we need to know
the date and time values the user has selected, not a mapped
representation in UTC.

This approach has also been chosen to be used by WHATWG[1] for
the `<input type="datetime-local">` element, where qualified
ISO8601 dates are explicitly forbidden, instead an unqualified
ISO8601 date is expected.
Adopting to this paradigm will allow to switch our datetime picker
to use native `<input type="datetime-local">` as some point
(Currently we're not happy with the Firefox UX and therefore stay with
a custom JavaScript based date picker, but it's good to be
prepared to switch by using a compatible format).

API changes
-----------

Also datetime offsets can now be provided, for DataHandler-as-an-API
usage, without the previously quite "unexpected" result that timezone
offsets were cropped.

That means, DataHandler now accepts five different *input* formats,
regardless of the persistence method of the respective field
(e.g. native datetime or timestamp field):

   * ISO8601 in LOCALTIME (Y-m-d\\TH:i:s):
     1999-11-11T11:11:11
     (interchange format between TYPO3 backend server/html/js)

   * ISO8601 with timezone offsets (Y-m-d\\TH:i:sP):
     1999-11-11T11:11:11Z
     1999-11-11T11:11:11:00:00
     1999-11-11T11:11:11+01:00
     (To be used as API)

   * SQL flavored date/times (Y-m-d H:i:s):
     1999-11-11 11:11:11
     1999-11-11
     11:11:11
     (To be used for copy operations of native datetime fields)

   * Unix timestamps (U):
     94231507
     (To be used for copy operations of unix timestamp fields)

   * DateTime objects with timezone offsets.
     NEW! Allows to pass-in `\DateTimeInterface` objects as-is without
     requring a to-string conversion.
     // same effect as ISO8601 LOCALTIME above
     new \DateTimeImmutable('1999-11-11T11:11:11')
     // same effect as ISO8601 with offset above
     new \DateTimeImmutable('1999-11-11T11:11:11+00:00')
     new \DateTimeImmutable('1999-11-11T11:11:11+01:00')
     // Offsets of custom timezones will be accepted an converted to
     // UTC for timestamp persistence and
     // LOCALTIME for native datetime fields
     // Heads UP: That means timezone offset is lost on persistence,
     //           but the date in time is stored correctly, without
     //           cropping offsets
     new \DateTimeImmutable(
         '1999-11-11T11:11:11',
         new \DateTimeZone('Asia/Jerusalem')
     )

This change is "breaking" because the previous fake ISO8601 UTC0 strings
will now be interpreted as *real* UTC0 and the offset to the server
timezone will be normalized instead of being cropped.

That means: For those who used DataHandler as an API, adaptions will be
needed to remove previous "workarounds" for DataHandlers broken DateTime
handling.

TODO:
 * It will make most sense to combine this with an optional timezone
   field that is made available from datetime fields and allows
   to store original timezone information, which then allows:
 * Switch native datetime fields from localtime to UTC and
   store desired localtime date and timezone information in a separate
   field together with timezone database version and possibly date of
   conversion
   (topic for another patch, needs migration wizard, field auto
   generation and so on…)
 * Streamline DatabaseRowDateTimeFields and DatetimeElement,
   the former needs to transform from DB to DateTime for all types
   (not just native datetime fields),
   DateTimeElement must not care about the persistence format,
   but expect a DateTime object and format that to an unqualified
   ISO8601 string for the backend form.

[1] https://en.wikipedia.org/wiki/ISO_8601#Local_time_(unqualified)
[2] https://html.spec.whatwg.org/multipage/common-microsyntaxes.html#local-dates-and-times
[3] https://html.spec.whatwg.org/multipage/input.html#local-date-and-time-state-(type=datetime-local)
[4] https://developer.mozilla.org/en-US/docs/Web/HTML/Element/input/datetime-local

Releases: main
Resolves: #105549
Related: #104309
Change-Id: Iaaf823ee57564e8835ac129707def927ae49804f