[WIP][TASK] Transform all dates in DatabaseRowDateTimeFields to DateTime
Instead of parsing only native datetime fields into ISO8601
and leaving timestamp fields as-is (to be parsed by DatetimeElement),
DatabaseRowDateTimeFields is now adapted to take responsibility for
consistent parsing of all type=datetime TCA fields and resolves
to a `DateTimeImmutable` object, which is then used by data
consumers like DatetimeElement and TcaRecordTitle.

This refactoring has revealed more edge cases where "0" was used as a
date string value instead of an ISO8601 string. This is now streamlined,
allowing to remove special treatment for this peculiarity in order to
cleanup form engine validation.

This refactoring also allows to streamline nullable timestamp handling,
were "0" was interpreted as an empty value, instead of
1970-01-01 00:00 UTC. The styleguide example inputdatetime_31 is adapted
to avoid the anti-pattern of default=0 for nullable datetime timestamp
fields, which was used as a workaround to opt-out from the formengine
range.lower validation by implicitly making a range-exception for the
default value. (Hint: a default value must adhere to any validation,
which is now fixed for nullable fields – for non-nullable fields the
behavior is kept as-is for compatibility reasons).

TODO:
 * share date string to DateTime conversion between
   DataHandler, DatabaseRowDateTimeFields
   (and ideally RecordFieldTransformer and extbase DataMapper)
   …this should probably already be prepared in #105549

Releases: main
Resolves: #105643
Resolves: #96890
Related: #105549
Related: #103693
Change-Id: I869b2aaff67d01b853465f873a229664501fcf97