[WIP][TASK] Transform all dates in DatabaseRowDateTimeFields to DateTime
Instead of parsing only native datetime fields into ISO8601
and leaving timestamp fields as-is (to be parsed by DatetimeElement),
DatabaseRowDateTimeFields is now adapted to take responsibility for
consistent parsing of all type=datetime TCA fields and resolves
to a `DateTimeImmutable` object, which is then used by data
consumers like DatetimeElement and TcaRecordTitle.

This refactoring has revealed more edge cases where "0" was used as a
date string value instead of an ISO8601 string. This is now streamlined,
allowing to remove special treatment for this peculiarity in order to
cleanup form engine validation.

TODO:
 * share date string to DateTime conversion between
   DataHandler, DatabaseRowDateTimeFields
   (and ideally RecordFieldTransformer and extbase DataMapper)
   …this should probably already be prepared in #105549

Releases: main
Resolves: #
Change-Id: I869b2aaff67d01b853465f873a229664501fcf97