[BUGFIX] Respect that native time fields are nullable by default
Since #97367 native datetime fields are generated as NULLABLE by
default. Honor this new default in DataHandler and FormEngine
in order to allow midnight – 00:00:00 – to be a valid value
of nullable native fields.

Also (re)allow to use non-nullable fields by settings false explicitly
in the TCA scema (for backwards compat for fields that were not
nullable in TYPO3 v12).

Note: 00:00:00 handling for time fields has been fixed in v12 by #97216,
but was effectively broken with #92900, which changed the reset value to
00:00:00 (which is the db-native empty value) instead of a value that is
intended to be understood by formengine (used to be timestamp integers –
as the original transformation target of DatabaseRowDateTimeFields, but
is ISO8601 nowadays).

Resolves: #105666
Related: #97367
Releases: main, 13.4
Change-Id: Idb27725f69554c040545984cac5690f2b0a34c21