[TASK] Modernize JavaScript and CSS buildchain
Use ECMAScript 2023 as TypeScript target
----------------------------------------

Use modern JavaScript targets for TypeScript and terser, and enable
terser `module` mode which allows to compress and mangle variables
and properties, based of the knowledge that the JavaScript are loaded
as modules (via `import`/`<script type="module">` and not as traditional
non-module `<script>` sources.

Replace clean-css with cssnano
------------------------------

Replace clean-css with cssnano, which is a native postcss plugin
that can operate on the postcss CSS AST directly, without the need
for an own CSS parser implementation.
This means that this plugins is less likely to miss support for new
CSS functionality like container queries (which clean-css – in the
version pulled in by postcss-clean – does not yet support).

Replace html-minifier with htmlnano
-----------------------------------

Replace the ancient library minify-html-literals with litnano (htmlnano
for lit template literals), as minify-html-literals pulls in insecure
dependencies and does not allow for asynchronous css processing
(needed for cssnano usage).

Allow generation of (temporary) JavaScript sourcemaps
-----------------------------------------------------

The TypeScript buildchain is adapted to support temporary generation of
sourcemaps, which may become helpful for debugging and development, now
that more variables are being mangled due to the activation of the
terser `module` option.

A sourcemap can be built via:

  npm run build-js -- --sourcemaps

  # Or only build sourcemaps for changed files
  npm run build -- watch --sourcemaps

Please note that the sourcemap is inlined into the source file
and needs to be rebuilt without --sourcemaps enabled in order
to be mergable. (CI will verify)

Asynchronous exection task
--------------------------

The TypeScirpt post-processing-copy task and the terser task are merged
into a new combined asynchronous grunt task, allowing to minify lit html
literals first and apply terser minification afterwards.
(Previously terser was executed first and minify-html-literals
afterwards, which worked because terser did not mangle `html`
literals without `module` mode being enabled)

The combined – and most importantly asynchronous – grunt task allows
to perform the switch to cssnano instead of clean-css,
which also requires asynchronous execution.

Terser is updated in order to use latest minification routines,
while rebuild most source files anyway:

Miscellaneous build optimizations
---------------------------------

 * Fully parallelize CPU-bound tasks
   (`npm run build` now takes ~18sec instead of ~40sec)
 * Remove concurrent wrappers for io-bound copy operations.
   Thread creation creates unneeded overhead, resulting in 200-400ms
   slowdowns instead of performance gains.
 * Use ES6 modules where possible
 * Move Build/util to Build/lib

Commands executed:
  npm remove \
    postcss-clean \
    rollup-plugin-postcss \
    minify-html-literals \
    es-module-lexer
  npm install --save-dev \
    terser@^5.39 \
    rollup@^4.36 \
    @rollup/plugin-commonjs@^28 \
    @rollup/plugin-node-resolve@^16 \
    postcss@^8.5 \
    cssnano@^7 \
    litnano@^1 \
    autoprefixer@^10.4.21 \
    grunt-sass@v4
  # Update to depend on cssnano v7 instead of v6
  npm install --save-dev @ckeditor/ckeditor5-dev-utils@^46
  npm audit fix
  grep -rl clean-css Build/Sources | xargs sed -i '/clean-css/d'
  npm run build

Resolves: #106357
Resolves: #106358
Releases: main, 13.4
Change-Id: Ibb667062b809d06b595e8144f149a0e67200235b