[TASK] Replace clean-css with cssnano
Replace clean-css with cssnano, which is a native postcss plugin
that can operate on the postcss CSS AST directly, without the need
for an own CSS parser implementation.
This means that this plugins is less likely to miss support for new
CSS functionality like container queries (which clean-css – in the
version pulled in by postcss-clean – does not yet support).

Since cssnano requires asynchronous execution, minify-html-literals
is wrapped to support an asynchronous css minification callback.

Commands executed:
  grep -rl clean-css Build/Sources | xargs sed -i '/clean-css/d'
  npm remove postcss-clean
  npm remove rollup-plugin-postcss
  npm install --save-dev cssnano
  # Update to depend on cssnano v7 instead of v6
  npm install --save-dev @ckeditor/ckeditor5-dev-utils@^46
  # Depend on postcss directly, as we use it's API in grunt and rollup
  npm install --save-dev postcss

Resolves: #106358
Releases: main, 13.4
Change-Id: If1a317ce3cc60ec6dc422da13c5fa42dca12d84d